Remove retry capabilities
The retry capability wasn't tested and I couldn't make it
works.

I think there was a bug, because the retry policy was
applied aroud the `(submit executor task)`, hence always
returning a future, even in case of exception and never
retrying ...

This feature may be reincorporated but should be implemented
inside the task. However this could also be a responsability
of the caller.